Use react hook forms and zod validation

+++ actions.ts
@@ -3,13 +3,19 @@
 import prisma from "./lib/prisma";
 import { revalidatePath } from "next/cache";
 import { fail, ok, sleep } from "./lib/utils";
-import { ActionResult, PetEssentials } from "./lib/types";
+import { ActionResult } from "./lib/types";
+import { petFormSchema, petIdSchema } from "./lib/validations";
 
-export async function addPet(pet: PetEssentials): Promise<ActionResult<null>> {
+export async function addPet(pet: unknown): Promise<ActionResult<null>> {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(pet);
+  if (validatedPet.success === false) {
+    return fail("Invalid pet data");
+  }
   try {
     await prisma.pet.create({
-      data: pet,
+      data: validatedPet.data,
     });
     revalidatePath("/app", "layout");
     return ok(null);
@@ -19,14 +25,20 @@ export async function addPet(pet: PetEssentials): Promise<ActionResult<null>> {
 }
 
 export async function editPet(
-  newPet: PetEssentials,
-  id: string
+  newPet: unknown,
+  id: unknown
 ): Promise<ActionResult<null>> {
   await sleep(1000);
+
+  const validatedId = petIdSchema.safeParse(id);
+  const validatedPet = petFormSchema.safeParse(newPet);
+  if (validatedPet.success === false || validatedId.success === false) {
+    return fail("Invalid pet data");
+  }
   try {
     await prisma.pet.update({
-      where: { id: id },
-      data: newPet,
+      where: { id: validatedId.data },
+      data: validatedPet.data,
     });
     revalidatePath("/app", "layout");
     return ok(null);
@@ -35,11 +47,18 @@ export async function editPet(
   }
 }
 
-export async function checkoutPet(id: string) {
+export async function checkoutPet(petId: unknown) {
   await sleep(1000);
+
+  const validatedId = petIdSchema.safeParse(petId);
+
+  if (validatedId.success === false) {
+    return fail("Invalid pet data");
+  }
+
   try {
     await prisma.pet.delete({
-      where: { id: id },
+      where: { id: validatedId.data },
     });
     revalidatePath("/app", "layout");
   } catch (error) {
